feat: Add Terragrunt configuration file
The  file is added to the projects directory. This file is responsible for generating the auto_provider.tf and terragrunt_backend.tf files, which are used for setting up provider configurations and remote state backend.

The  file enables a centralized configuration and allows for inheritance of settings across different environments.